requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv
psycopg2-binary